requirements.txt
Pillow==10.0.1
sphinx==7.2.6
python-magic-bin==0.4.14
requests==2.31.0
build==1.0.3
twine==4.0.2